+++ RELEASE_NOTES.md
@@ -12,7 +12,7 @@ _TBD_
   * Updated reports to display changes
   * Display / allow edit of c3/c4 for 2015 product CQMS
 * Updated links in /overview
-* Integrated with changes to /developers call
+* Integrated view/edit developers with changes to /developers call
 
 ---